make system dictionary singleton
