+++ requirements-dev.txt
@@ -10,3 +10,4 @@ typed-ast==1.4.1
 typing-extensions==3.7.4.2
 twine==3.2.0
 wheel==0.35.1
+psutil==5.9.0
