+++ tests/test_dic.py
@@ -15,10 +15,8 @@
 import os
 import sys
 import unittest
-from janome.sysdic import all_fstdata, entries, mmap_entries, connections, chardef, unknowns
+from janome.sysdic import connections
 from janome.dic import (
-    SystemDictionary,
-    MMapSystemDictionary,
     UserDictionary,
     CompiledUserDictionary,
     FILE_USER_FST_DATA,
@@ -32,78 +30,8 @@ parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 sys.path.insert(0, parent_dir)
 
 
-class TestDictionary(unittest.TestCase):
-    def test_system_dictionary_ipadic(self):
-        matcher = Matcher(all_fstdata())
-        sys_dic = SystemDictionary(entries(), connections, chardef.DATA, unknowns.DATA)
-        self.assertEqual(7, len(sys_dic.lookup('形態素'.encode('utf-8'), matcher)))
-        self.assertEqual(1, sys_dic.get_trans_cost(0, 1))
-        self.assertEqual({'HIRAGANA': []}, sys_dic.get_char_categories('は'))
-        self.assertEqual({'KATAKANA': []}, sys_dic.get_char_categories('ハ'))
-        self.assertEqual({'KATAKANA': []}, sys_dic.get_char_categories('ﾊ'))
-        self.assertEqual({'KANJI': []}, sys_dic.get_char_categories('葉'))
-        self.assertEqual({'ALPHA': []}, sys_dic.get_char_categories('C'))
-        self.assertEqual({'ALPHA': []}, sys_dic.get_char_categories('Ｃ'))
-        self.assertEqual({'SYMBOL': []}, sys_dic.get_char_categories('#'))
-        self.assertEqual({'SYMBOL': []}, sys_dic.get_char_categories('＃'))
-        self.assertEqual({'NUMERIC': []}, sys_dic.get_char_categories('5'))
-        self.assertEqual({'NUMERIC': []}, sys_dic.get_char_categories('５'))
-        self.assertEqual({'KANJI': [], 'KANJINUMERIC': ['KANJI']}, sys_dic.get_char_categories('五'))
-        self.assertEqual({'GREEK': []}, sys_dic.get_char_categories('Γ'))
-        self.assertEqual({'CYRILLIC': []}, sys_dic.get_char_categories('Б'))
-        self.assertEqual({'DEFAULT': []}, sys_dic.get_char_categories('𠮷'))
-        self.assertEqual({'DEFAULT': []}, sys_dic.get_char_categories('한'))
-        self.assertTrue(sys_dic.unknown_invoked_always('ALPHA'))
-        self.assertFalse(sys_dic.unknown_invoked_always('KANJI'))
-        self.assertTrue(sys_dic.unknown_grouping('NUMERIC'))
-        self.assertFalse(sys_dic.unknown_grouping('KANJI'))
-        self.assertEqual(2, sys_dic.unknown_length('HIRAGANA'))
-
+class TestUserDictionary(unittest.TestCase):
     def test_property_types(self):
-        matcher = Matcher(all_fstdata())
-        sys_dic = SystemDictionary(entries(), connections, chardef.DATA, unknowns.DATA)
-        # entry in the system dictionary
-        entry = sys_dic.lookup('すもも'.encode('utf8'), matcher)[0]
-        self.assertTrue(type(entry[1]) is str)
-        self.assertTrue(type(entry[0]) is int)
-        self.assertTrue(type(entry[2]) is int)
-        self.assertTrue(type(entry[3]) is int)
-        self.assertTrue(type(entry[4]) is int)
-
-        entry_extra = sys_dic.lookup_extra(entry[0])
-        self.assertTrue(type(entry_extra[0]) is str)
-        self.assertTrue(type(entry_extra[1]) is str)
-        self.assertTrue(type(entry_extra[2]) is str)
-        self.assertTrue(type(entry_extra[3]) is str)
-        self.assertTrue(type(entry_extra[4]) is str)
-        self.assertTrue(type(entry_extra[5]) is str)
-
-        # unknown entry
-        entry = sys_dic.unknowns.get(u'HIRAGANA')[0]
-        self.assertTrue(type(entry[3]) is str)
-        self.assertTrue(type(entry[0]) is int)
-        self.assertTrue(type(entry[1]) is int)
-        self.assertTrue(type(entry[2]) is int)
-
-        # mmap dict etnry
-        matcher = Matcher(all_fstdata())
-        mmap_dic = MMapSystemDictionary(mmap_entries(), connections, chardef.DATA, unknowns.DATA)
-        entry = mmap_dic.lookup(u'すもも'.encode('utf8'), matcher)[0]
-        self.assertTrue(type(entry[1]) is str)
-        self.assertTrue(type(entry[0]) is int)
-        self.assertTrue(type(entry[2]) is int)
-        self.assertTrue(type(entry[3]) is int)
-        self.assertTrue(type(entry[4]) is int)
-
-        entry_extra = mmap_dic.lookup_extra(entry[0])
-        self.assertTrue(type(entry_extra[0]) is str)
-        self.assertTrue(type(entry_extra[1]) is str)
-        self.assertTrue(type(entry_extra[2]) is str)
-        self.assertTrue(type(entry_extra[3]) is str)
-        self.assertTrue(type(entry_extra[4]) is str)
-        self.assertTrue(type(entry_extra[5]) is str)
-
-        # entry in the user defined dictionary
         user_dic = UserDictionary(user_dict=os.path.join(parent_dir, 'tests/user_ipadic.csv'),
                                   enc='utf8', type='ipadic', connections=connections)
         entry = user_dic.lookup('東京スカイツリー'.encode('utf8'))[0]
@@ -112,19 +40,6 @@ class TestDictionary(unittest.TestCase):
         self.assertTrue(type(entry[2]) is int)
         self.assertTrue(type(entry[3]) is int)
         self.assertTrue(type(entry[4]) is int)
-
-    def test_system_dictionary_cache(self):
-        matcher = Matcher(all_fstdata())
-        sys_dic = SystemDictionary(entries(), connections, chardef.DATA, unknowns.DATA)
-        self.assertEqual(11, len(sys_dic.lookup('小書き'.encode('utf8'), matcher)))
-        self.assertEqual(11, len(sys_dic.lookup('小書き'.encode('utf8'), matcher)))
-        self.assertEqual(11, len(sys_dic.lookup('小書きにしました'.encode('utf8'), matcher)))
-
-        self.assertEqual(10, len(sys_dic.lookup('みんなと'.encode('utf8'), matcher)))
-        self.assertEqual(10, len(sys_dic.lookup('みんなと'.encode('utf8'), matcher)))
-
-        self.assertEqual(2, len(sys_dic.lookup('叩く'.encode('utf8'), matcher)))
-        self.assertEqual(2, len(sys_dic.lookup('叩く'.encode('utf8'), matcher)))
 
     def test_user_dictionary(self):
         # create user dictionary from csv
